conditional benchmrking

+++ payload_uppaal/process_payload/main.sh
@@ -6,14 +6,19 @@ OUTPUT_DIR=$2
 cp run.sh $INPUT_DIR/run.sh
 RUN_DIR=`cd "$(dirname "$0")" && pwd`
 
-exec_start_time=`date +"%Y-%m-%d %H:%M:%S.%3N"`
-sed -i "s/EXEC_START_TIME/$exec_start_time/" ./timedata.txt
+if [ -e timedata.txt ]
+then
+    exec_start_time=`date +"%Y-%m-%d %H:%M:%S.%3N"`
+    sed -i "s/EXEC_START_TIME/$exec_start_time/" ./timedata.txt
+fi 
 
 sh $RUN_DIR/run.sh $@
 
-exec_end_time=`date +"%Y-%m-%d %H:%M:%S.%3N"`
-sed -i "s/EXEC_END_TIME/$exec_end_time/" ./timedata.txt
-cp ./timedata.txt $OUTPUT_DIR/timedata.txt
-
+if [ -e timedata.txt ]
+then
+    exec_end_time=`date +"%Y-%m-%d %H:%M:%S.%3N"`
+    sed -i "s/EXEC_END_TIME/$exec_end_time/" ./timedata.txt
+    cp ./timedata.txt $OUTPUT_DIR/timedata.txt
+fi 
 
 # --- EOF ---
